fix: write rom cast

+++ main.c
@@ -5052,10 +5052,6 @@ void write_ram_ds1307(void)
    PCF8583_write_byte(ds1307_Min_Save_To, phut_save_to);
    //=======
 
-   // PCF8583_write_byte(ds1307_input_dc_low, make8((unsigned long)(input_dc_low * 10), 0));
-   // PCF8583_write_byte(ds1307_input_dc_low + 1, make8((unsigned long)(input_dc_low * 10), 1));
-   // PCF8583_write_byte(ds1307_delta_dc, make8((unsigned long)(delta_dc * 10), 0));
-   // PCF8583_write_byte(ds1307_delta_dc + 1, make8((unsigned long)(delta_dc * 10), 1));
    PCF8583_write_byte(ds1307_input_dc_low, (unsigned long)(input_dc_low * 10));
    PCF8583_write_byte(ds1307_delta_dc, (unsigned long)(delta_dc * 10));
    PCF8583_write_byte(ds1307_flag_accu_error_save_log, flag_accu_error_save_log);
@@ -5147,8 +5143,8 @@ void write_data_ee(void)
    wee(ee_Min_Save_To, phut_save_to);
    //=======
 
-   wee16(ee_input_dc_low, (input_dc_low * 10));
-   wee16(ee_delta_dc, (delta_dc * 10));
+   wee16(ee_input_dc_low, (unsigned long)(input_dc_low * 10));
+   wee16(ee_delta_dc, (unsigned long)(delta_dc * 10));
    wee(ee_flag_accu_error_save_log, flag_accu_error_save_log);
 }
